[UserTimingL3] Always serialize detail
UserTimingL3 specs that the detail property of the measure method and
the performance mark constructor both need to serialize the detail
property:

In https://w3c.github.io/user-timing/#measure-method:
"Set entry's detail attribute as follows:
If startOrMeasureOptions is present and is a
PerformanceMeasureOptions object, set it to the result of calling
the StructuredSerialize algorithm on startOrMeasureOptions's detail.
Otherwise, set it as null."

In https://w3c.github.io/user-timing/#the-performancemark-constructor:
"Set entry's detail attribute to the result of calling the
StructuredSerialize algorithm on markOptions's detail."

In the current implementation of the performancemark constructor, it
serializes the detail only when the world of constructor and
the world of accessing the detail are different. To correct this, we
should serialize in both situations.

In the current implementation of the measure method, it does serialize
in both cases, but it swallows the exception. We should not swallow the
exception so as to "fail early, fail loudly".

Bug:953857, 963926
Change-Id: I471c4d207d9c05bc9b9b8b65315c49b0aa73a672