Freezing on load.
In change https://crrev.com/0daefdcf34ba9a8c21ab907b2159c5652169a45b
caused a dcheck failure occasionally because of an implicit open of a
document inside an implicit close while running the
wpt/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/abort-while-navigating.window.html
wpt test.

See: https://chromium-swarm.appspot.com/task?id=44fb7387f8411e10
for the failure.

https://chromium-review.googlesource.com/c/chromium/src/+/1628499
reverted the original code so this is a reland of some of that code.

Because of the complexity of that value use a simpler
approach and use the Frame::IsLoading attribute instead which doesn't
exhibit these cycles.

BUG=907125

Change-Id: Ifb805cf8ab90b6a228f47d88cb2ee872d9a21239